requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
adal==1.2.7
argcomplete==1.12.3
azure-common==1.1.28
azure-core==1.21.1
azure-graphrbac==0.61.1
azure-identity==1.7.0
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==8.2.0
azure-mgmt-core==1.3.0
azure-mgmt-keyvault==9.3.0
azure-mgmt-resource==20.1.0
azure-mgmt-storage==19.1.0
azureml-core==1.39.0
azureml-dataprep==2.27.1
azureml-dataprep-native==38.0.0
azureml-dataprep-rslex==2.3.1
azureml-dataset-runtime==1.39.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cloudpickle==2.0.0
contextlib2==21.6.0
cryptography==36.0.1
distro==1.7.0
docker==5.0.3
dotnetcore2==2.1.23
humanfriendly==10.0
idna==3.3
isodate==0.6.1
jeepney==0.7.1
jmespath==0.10.0
jsonpickle==2.1.0
knack==0.9.0
msal==1.17.0
msal-extensions==0.3.1
msrest==0.6.21
msrestazure==0.6.4
ndg-httpsclient==0.5.1
numpy==1.22.2
oauthlib==3.2.0
packaging==21.3
pandas==1.4.1
paramiko==2.9.2
pathspec==0.9.0
pkginfo==1.8.2
portalocker==2.4.0
pyarrow==3.0.0
pyasn1==0.4.8
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==3.0.7
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
#pywin32==227
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
SecretStorage==3.3.1
six==1.16.0
tabulate==0.8.9
urllib3==1.26.7
websocket-client==1.3.1
